renamed and refactored CustomConnector into ParameterizableHardwareConnector
