+++ tests/utils/connector.py
@@ -7,9 +7,41 @@ from streamflow.core.data import StreamWrapperContextManager
 from streamflow.core.deployment import (
     Connector,
     Location,
+    LOCAL_LOCATION,
 )
-from streamflow.core.scheduling import AvailableLocation
+from streamflow.core.scheduling import AvailableLocation, Hardware
+from streamflow.deployment.connector import LocalConnector
 from streamflow.log_handler import logger
+
+
+class ParameterizableHardwareConnector(LocalConnector):
+    def __init__(
+        self,
+        deployment_name: str,
+        config_dir: str,
+        hardware: Hardware,
+        transferBufferSize: int = 2**16,
+    ):
+        super().__init__(deployment_name, config_dir, transferBufferSize)
+        self.hardware = hardware
+
+    async def get_available_locations(
+        self,
+        service: str | None = None,
+        input_directory: str | None = None,
+        output_directory: str | None = None,
+        tmp_directory: str | None = None,
+    ) -> MutableMapping[str, AvailableLocation]:
+        return {
+            LOCAL_LOCATION: AvailableLocation(
+                name=LOCAL_LOCATION,
+                deployment=self.deployment_name,
+                service=service,
+                hostname="localhost",
+                slots=1,
+                hardware=self.hardware,
+            )
+        }
 
 
 class FailureConnectorException(Exception):
