+++ tests/utils/deployment.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import asyncio
 import os
 import tempfile
-from typing import cast
+from typing import cast, MutableSequence
 
 import asyncssh
 import asyncssh.public_key
@@ -17,7 +17,10 @@ from streamflow.core.deployment import (
     LOCAL_LOCATION,
     Location,
     WrapsConfig,
+    BindingFilter,
+    Target,
 )
+from streamflow.core.workflow import Job
 from streamflow.deployment import DefaultDeploymentManager
 from tests.utils.data import get_data_path
 
@@ -231,3 +234,14 @@ async def get_ssh_deployment_config(_context: StreamFlowContext):
         external=False,
         lazy=False,
     )
+
+
+class ReverseTargetsBindingFilter(BindingFilter):
+    async def get_targets(
+        self, job: Job, targets: MutableSequence[Target]
+    ) -> MutableSequence[Target]:
+        return targets[::-1]
+
+    @classmethod
+    def get_schema(cls) -> str:
+        return ""
